ESM conversion  (#61)
* initial pass at updating packages

* nest is starting up

* fixed lint warnings

* updated travis

* version update

* nestjs version upgrade; travis config

* updated typescript version; target; nest-cli

* made imports .js

* additional updates to try + get nest esm working

* exploring moving to native node aliases

* revamped project setup based on another nestjs template using esm

* updated matrix client sdk verison

* tidy up

* removed usage of .js suffix

* updating travis setup

* updated travis to use pnpm

* updates to have jest config file mjs

* replaced jest with vitest

* updated test run by travis

* tidy up scripts

* Fixes formatting of MessagePattern definitions

Corrects the formatting of `@MessagePattern` definitions in the controller.

Improves code readability and consistency by aligning the `cmd` property with the event type from the `alkemioMatrixAdapterLib`.

* Migrates to pnpm and updates dependencies.

Updates the Dockerfile to use pnpm instead of npm for package management, which improves dependency installation speed and efficiency.

Also updates the base Node.js image version and adds necessary configuration files for production builds.

* Downgrades version and configures ts

Downgrades the package version to align with
the release process.

Updates the tsconfig to improve type checking and
code consistency.

---------

Co-authored-by: Valentin Yanakiev <valentin.yanakiev@gmail.com>